Extract utility function to new util/misc module
